+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/PhoneAuthCodeAutoRetrievalTimeOutEvent.java
@@ -2,8 +2,8 @@ package com.androidhuman.rxfirebase2.auth;
 
 import com.google.auto.value.AutoValue;
 
-import android.support.annotation.CheckResult;
-import android.support.annotation.NonNull;
+import androidx.annotation.CheckResult;
+import androidx.annotation.NonNull;
 
 @AutoValue
 public abstract class PhoneAuthCodeAutoRetrievalTimeOutEvent extends PhoneAuthEvent {