+++ firebase-database/src/main/java/com/androidhuman/rxfirebase2/database/UpdateChildrenObserver.java
@@ -5,10 +5,9 @@ import com.google.firebase.database.DatabaseReference;
 
 import com.androidhuman.rxfirebase2.core.SimpleDisposable;
 
-import android.support.annotation.NonNull;
-
 import java.util.Map;
 
+import androidx.annotation.NonNull;
 import io.reactivex.Completable;
 import io.reactivex.CompletableObserver;
 
@@ -40,7 +39,8 @@ final class UpdateChildrenObserver extends Completable {
         }
 
         @Override
-        public void onComplete(DatabaseError databaseError, DatabaseReference databaseReference) {
+        public void onComplete(DatabaseError databaseError,
+                @NonNull DatabaseReference databaseReference) {
             if (!isDisposed()) {
                 if (null != databaseError) {
                     observer.onError(databaseError.toException());