+++ firebase-database/src/main/java/com/androidhuman/rxfirebase2/database/model/Some.java
@@ -2,7 +2,7 @@ package com.androidhuman.rxfirebase2.database.model;
 
 import com.google.auto.value.AutoValue;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 @AutoValue
 public abstract class Some<T> extends DataValue<T> {