+++ firebase-firestore/src/main/java/com/androidhuman/rxfirebase2/firestore/DeleteDocumentObserver.java
@@ -5,8 +5,7 @@ import com.google.firebase.firestore.DocumentReference;
 
 import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
 
-import android.support.annotation.NonNull;
-
+import androidx.annotation.NonNull;
 import io.reactivex.Completable;
 import io.reactivex.CompletableObserver;