+++ firebase-firestore/src/main/java/com/androidhuman/rxfirebase2/firestore/DocumentObserver.java
@@ -8,9 +8,8 @@ import com.google.firebase.firestore.Source;
 import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
 import com.androidhuman.rxfirebase2.firestore.model.Value;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import io.reactivex.Single;
 import io.reactivex.SingleObserver;