+++ firebase-storage/src/main/java/com/androidhuman/rxfirebase2/storage/PutFileWithMetadataObserver.java
@@ -8,8 +8,8 @@ import com.google.firebase.storage.UploadTask;
 import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
 
+import androidx.annotation.NonNull;
 import io.reactivex.Single;
 import io.reactivex.SingleObserver;